requirements.txt
absl-py==2.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
dronekit==2.9.2
flatbuffers==20181003210633
future==1.0.0
gcloud==0.18.3
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
jws==0.1.3
lxml==5.1.0
MAVProxy==1.8.70
monotonic==1.6
numpy==1.26.4
oauth2client==4.1.3
opencv-contrib-python==3.4.13.47
pillow==10.2.0
pkg_resources==0.0.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pymavlink==2.4.41
pyparsing==3.1.2
Pyrebase4==4.8.0
pyserial==3.5
python-jwt==2.0.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sounddevice==0.4.6
tflite-runtime @ https://github.com/google-coral/pycoral/releases/download/v2.0.0/tflite_runtime-2.5.0.post1-cp39-cp39-linux_aarch64.whl
tflite-support==0.4.3
urllib3==1.26.18
